+++ modules-services/service-templates/src/main/java/org/openprovenance/prov/service/readers/JsonOrCsvMessageBodyReader.java
@@ -1,4 +1,4 @@
-package org.openprovenance.prov.service;
+package org.openprovenance.prov.service.readers;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;