+++ modules-services/service-templates/src/main/java/org/openprovenance/prov/service/readers/TableKey.java
@@ -1,4 +1,4 @@
-package org.openprovenance.prov.service;
+package org.openprovenance.prov.service.readers;
 
 public class TableKey {
     public String isA;