+++ RELEASES.md
@@ -25,7 +25,7 @@ See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 * service-templates: navigation tab
 * prov-template-compiler: getting level0 types for template variables
 * service-templates: displaying level0 types colours for template variables
-
+* service-templates: viewing template instance value by property
 
 
 `2.0.5`